refactor: rename and consolidate CreateOnlyBool attribute plan modifiers for clarity
